+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-* 2022.01.12    Current
+* 2022.01.13    Current
+  * `network.c`: `inet2str()` Returns IP:PORT
+  * `inifile.c`: Many minor logfile tweaks
   * `ts-warp.c`: `read_ini()` added more allowed characters in sections: a-zA-Z0-9_\t -+()
   * `inifile.c`, `pidlist.c`: `show_ini()`, `pidlist_show()` logs using `LOG_CRIT`